Lung_APP/requirements.txt
streamlit
tensorflow-cpu
opencv-python-headless
matplotlib
pillow
numpy